requirements.txt
Django==1.6.1
Markdown==2.3.1
South==0.8.4
dj-database-url==0.2.2
dj-static==0.0.5
django-filter==0.7
djangorestframework==2.3.10
gunicorn==18.0
newrelic==2.8.0.7
psycopg2==2.5.1
static==0.4
wsgiref==0.1.2
